More changes for spring boot conversion.  Unit tests, modules, some servlets.

+++ src/main/java/org/oscarehr/common/dao/AbstractDao.java
@@ -142,7 +142,7 @@ public abstract class AbstractDao<T extends AbstractModel<?>> {
 			MiscUtils.getLogger().error("Criteria search count returned null result");
 			result = -1;
 		}
-		return ((Integer)result);
+		return ((Long)result).intValue();
 	}
 	
 	protected int getMaxSelectSize() {
